refactor: align package layout with the architecture
Move import binding and canonical symbols into semantic/, HIR-to-PyIR lowering
into ir/, and expose frontend.build_hir as the only public frontend entry point
so parser objects never leave the frontend.
