--- src/coretrace_python/frontend/symbols.py
@@ -1,5 +1,0 @@
-"""Frontend support for canonical Python symbols."""
-
-from coretrace_python.ir.symbol import SymbolId
-
-__all__ = ["SymbolId"]
