+++ src/coretrace_python/ir/__init__.py
@@ -1,6 +1,5 @@
 """Python-aware intermediate representation."""
 
 from coretrace_python.ir.model import BasicBlock, FunctionIR, ModuleIR, Value
-from coretrace_python.source import SourceSpan
 
-__all__ = ["BasicBlock", "FunctionIR", "ModuleIR", "SourceSpan", "Value"]
+__all__ = ["BasicBlock", "FunctionIR", "ModuleIR", "Value"]
