+++ src/coretrace_python/semantic/__init__.py
@@ -1,0 +1,1 @@
+"""Semantic analyses computed from PyHIR: imports, symbols and lexical scopes."""
