+++ src/coretrace_python/semantic/symbols.py
